Separated container and vault tests + session records
