+++ src/ResourceManager/AzureBackup/Commands.AzureBackup.Test/SessionRecords/Microsoft.Azure.Commands.AzureBackup.Test.ScenarioTests.AzureBackupContainerTests/GetAzureBackupContainerTests.json
@@ -31,29 +31,29 @@
           "no-cache"
         ],
         "x-ms-request-id": [
-          "cb36f6ae-187d-4a30-b367-d714c0733647"
+          "ce4dfbf6-9708-49f2-9532-0dbc9e912bf7"
         ],
         "x-ms-client-request-id": [
-          "4b55cb3b-c4db-4dda-8db9-891256be4264",
-          "4b55cb3b-c4db-4dda-8db9-891256be4264"
+          "0482c73c-7196-49f8-96a0-3dcd0ccdaa90",
+          "0482c73c-7196-49f8-96a0-3dcd0ccdaa90"
         ],
         "Strict-Transport-Security": [
           "max-age=31536000; includeSubDomains"
         ],
         "x-ms-ratelimit-remaining-subscription-reads": [
-          "14999"
+          "14997"
         ],
         "x-ms-correlation-request-id": [
-          "cb36f6ae-187d-4a30-b367-d714c0733647"
+          "ce4dfbf6-9708-49f2-9532-0dbc9e912bf7"
         ],
         "x-ms-routing-request-id": [
-          "CENTRALUS:20150616T144830Z:cb36f6ae-187d-4a30-b367-d714c0733647"
+          "CENTRALUS:20150617T170930Z:ce4dfbf6-9708-49f2-9532-0dbc9e912bf7"
         ],
         "Cache-Control": [
           "no-cache"
         ],
         "Date": [
-          "Tue, 16 Jun 2015 14:48:30 GMT"
+          "Wed, 17 Jun 2015 17:09:30 GMT"
         ],
         "Server": [
           "Microsoft-IIS/8.0"
