--- src/ResourceManager/AzureBackup/Commands.AzureBackup.Test/SessionRecords/Microsoft.Azure.Commands.AzureBackup.Test.ScenarioTests.AzureBackupTests/GetAzureBackupVaultCredentialsTests.json
@@ -1,75 +1,0 @@
-{
-  "Entries": [
-    {
-      "RequestUri": "/Subscriptions/f5303a0b-fae4-4cdb-b44d-0e4c032dde26/resourceGroups/backuprg/providers/Microsoft.Backupseadev01/BackupVault/backuprn/certificates/IdMgmtInternalCert?api-version=2015-03-15",
-      "EncodedRequestUri": "L1N1YnNjcmlwdGlvbnMvZjUzMDNhMGItZmFlNC00Y2RiLWI0NGQtMGU0YzAzMmRkZTI2L3Jlc291cmNlR3JvdXBzL2JhY2t1cHJnL3Byb3ZpZGVycy9NaWNyb3NvZnQuQmFja3Vwc2VhZGV2MDEvQmFja3VwVmF1bHQvYmFja3Vwcm4vY2VydGlmaWNhdGVzL0lkTWdtdEludGVybmFsQ2VydD9hcGktdmVyc2lvbj0yMDE1LTAzLTE1",
-      "RequestMethod": "PUT",
-      "RequestBody": "{\r\n  \"properties\": {\r\n    \"certificate\": \"MIIC3TCCAcWgAwIBAgIQfAPxy3CY6JlCIBZLq9ZwVjANBgkqhkiG9w0BAQUFADAeMRwwGgYDVQQDExNXaW5kb3dzIEF6dXJlIFRvb2xzMB4XDTE1MDYxNjE1NDQ1MVoXDTE1MDYxODE1NTQ1MVowHjEcMBoGA1UEAxMTV2luZG93cyBBenVyZSBUb29sczCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJcmOVXMwQS3uiV3xRbeRbbQxjbS/eT/n3m7b9nd2NoPsMhGT+uvypAL3iG2wxd7HzqjZswIVMhsLqYMzQIv6kcAlbyosuiaT56ga/xLBZJGJg7DHA4ndQf4UnigmqLCZeE5dyvJ/U070iHEBS1M9pa6TPmYgIEVKFZ1p3qDnmHsz6DPb39vdzxpKAsseGD2ocdLmb0Eq1s7ZnS9djzLkiY7x2P715Iivz1PZx/JcjTAwCorfWU9elI+sdSY62DZKhZbrXnK+qFEYnRO1cPWXe9AvJynQvS+FRkD8HCLHweWXEvCs/BX8CGJbKi9XYnuDlhuxm4dLBCs92tXrhu2hUsCAwEAAaMXMBUwEwYDVR0lBAwwCgYIKwYBBQUHAwIwDQYJKoZIhvcNAQEFBQADggEBAGj+8Vo4WUUui7GepTrsRAKAx2ZpdfCTTBhsdQIxbtaz312xX93i6X0CCmJLRjeXNNoVv5b0zDFnIdoTqn3eraDNonkMUqNGGleC4gMnOLzbKURMWeJ2cfGIVGtHkSv8ZFgU+ttZtCPX6SpzAWcplCeEX8DNs87e9ew4jZg4gPYu9zEO+Z4iYQlGd55Pd8JiiCsVcs9KhtrbBHC9J4/tTAZsONPLqo3PFdXMqmuN5Yct+J5T93kW3FWb/X5oiNZXtn9O4j7T7rYSHje9IZnSoiNW0Nc9/JgQ5VPbUMpc6U6tE53yIQ1aWpLj49L+gxSoZ7KWefN/NDMGSXTzTuXJ9R0=\"\r\n  }\r\n}",
-      "RequestHeaders": {
-        "Content-Type": [
-          "application/json"
-        ],
-        "Content-Length": [
-          "1035"
-        ],
-        "Accept-Language": [
-          "en-us"
-        ],
-        "x-ms-version": [
-          "2013-03-01"
-        ],
-        "User-Agent": [
-          "Microsoft.Azure.Management.BackupServices.BackupServicesManagementClient/0.0.0.0"
-        ]
-      },
-      "ResponseBody": "{\r\n  \"properties\": {\r\n    \"certificate\": \"MIIC3TCCAcWgAwIBAgIQfAPxy3CY6JlCIBZLq9ZwVjANBgkqhkiG9w0BAQUFADAeMRwwGgYDVQQDExNXaW5kb3dzIEF6dXJlIFRvb2xzMB4XDTE1MDYxNjE1NDQ1MVoXDTE1MDYxODE1NTQ1MVowHjEcMBoGA1UEAxMTV2luZG93cyBBenVyZSBUb29sczCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJcmOVXMwQS3uiV3xRbeRbbQxjbS/eT/n3m7b9nd2NoPsMhGT+uvypAL3iG2wxd7HzqjZswIVMhsLqYMzQIv6kcAlbyosuiaT56ga/xLBZJGJg7DHA4ndQf4UnigmqLCZeE5dyvJ/U070iHEBS1M9pa6TPmYgIEVKFZ1p3qDnmHsz6DPb39vdzxpKAsseGD2ocdLmb0Eq1s7ZnS9djzLkiY7x2P715Iivz1PZx/JcjTAwCorfWU9elI+sdSY62DZKhZbrXnK+qFEYnRO1cPWXe9AvJynQvS+FRkD8HCLHweWXEvCs/BX8CGJbKi9XYnuDlhuxm4dLBCs92tXrhu2hUsCAwEAAaMXMBUwEwYDVR0lBAwwCgYIKwYBBQUHAwIwDQYJKoZIhvcNAQEFBQADggEBAGj+8Vo4WUUui7GepTrsRAKAx2ZpdfCTTBhsdQIxbtaz312xX93i6X0CCmJLRjeXNNoVv5b0zDFnIdoTqn3eraDNonkMUqNGGleC4gMnOLzbKURMWeJ2cfGIVGtHkSv8ZFgU+ttZtCPX6SpzAWcplCeEX8DNs87e9ew4jZg4gPYu9zEO+Z4iYQlGd55Pd8JiiCsVcs9KhtrbBHC9J4/tTAZsONPLqo3PFdXMqmuN5Yct+J5T93kW3FWb/X5oiNZXtn9O4j7T7rYSHje9IZnSoiNW0Nc9/JgQ5VPbUMpc6U6tE53yIQ1aWpLj49L+gxSoZ7KWefN/NDMGSXTzTuXJ9R0=\",\r\n    \"resourceId\": 3420983535961922368,\r\n    \"globalAcsNamespace\": \"seadev01rp1users\",\r\n    \"globalAcsHostName\": \"accesscontrol.windows.net\",\r\n    \"globalAcsRPRealm\": \"http://windowscloudbackup/m3\",\r\n    \"subject\": \"CN=Windows Azure Tools\",\r\n    \"validFrom\": \"2015-06-16T21:14:51+05:30\",\r\n    \"validTo\": \"2015-06-18T21:24:51+05:30\",\r\n    \"thumbprint\": \"3C6F6F4C967A814B373E16F4E21903279E068704\",\r\n    \"friendlyName\": \"\",\r\n    \"issuer\": \"CN=Windows Azure Tools\"\r\n  }\r\n}",
-      "ResponseHeaders": {
-        "Content-Length": [
-          "1409"
-        ],
-        "Content-Type": [
-          "application/json"
-        ],
-        "Expires": [
-          "-1"
-        ],
-        "Pragma": [
-          "no-cache"
-        ],
-        "x-ms-request-id": [
-          "8511ee99-3e25-405e-aa6f-0cb873bdab7c"
-        ],
-        "x-ms-client-request-id": [
-          "0c51abcf-3be9-4c40-ba65-de8e5ed0d233",
-          "0c51abcf-3be9-4c40-ba65-de8e5ed0d233"
-        ],
-        "Strict-Transport-Security": [
-          "max-age=31536000; includeSubDomains"
-        ],
-        "x-ms-ratelimit-remaining-subscription-writes": [
-          "1197"
-        ],
-        "x-ms-correlation-request-id": [
-          "8511ee99-3e25-405e-aa6f-0cb873bdab7c"
-        ],
-        "x-ms-routing-request-id": [
-          "CENTRALUS:20150616T155458Z:8511ee99-3e25-405e-aa6f-0cb873bdab7c"
-        ],
-        "Cache-Control": [
-          "no-cache"
-        ],
-        "Date": [
-          "Tue, 16 Jun 2015 15:54:58 GMT"
-        ],
-        "Server": [
-          "Microsoft-IIS/8.0"
-        ]
-      },
-      "StatusCode": 200
-    }
-  ],
-  "Names": {},
-  "Variables": {
-    "SubscriptionId": "f5303a0b-fae4-4cdb-b44d-0e4c032dde26"
-  }
-}
