+++ src/ResourceManager/AzureBackup/Commands.AzureBackup.Test/SessionRecords/Microsoft.Azure.Commands.AzureBackup.Test.ScenarioTests.AzureBackupVaultTests/GetAzureBackupVaultCredentialsTests.json
@@ -1,0 +1,75 @@
+{
+  "Entries": [
+    {
+      "RequestUri": "/Subscriptions/f5303a0b-fae4-4cdb-b44d-0e4c032dde26/resourceGroups/backuprg/providers/Microsoft.Backupseadev01/BackupVault/backuprn/certificates/IdMgmtInternalCert?api-version=2015-03-15",
+      "EncodedRequestUri": "L1N1YnNjcmlwdGlvbnMvZjUzMDNhMGItZmFlNC00Y2RiLWI0NGQtMGU0YzAzMmRkZTI2L3Jlc291cmNlR3JvdXBzL2JhY2t1cHJnL3Byb3ZpZGVycy9NaWNyb3NvZnQuQmFja3Vwc2VhZGV2MDEvQmFja3VwVmF1bHQvYmFja3Vwcm4vY2VydGlmaWNhdGVzL0lkTWdtdEludGVybmFsQ2VydD9hcGktdmVyc2lvbj0yMDE1LTAzLTE1",
+      "RequestMethod": "PUT",
+      "RequestBody": "{\r\n  \"properties\": {\r\n    \"certificate\": \"MIIC3TCCAcWgAwIBAgIQaa4WR+3Ra7pPjIzudnuuQTANBgkqhkiG9w0BAQUFADAeMRwwGgYDVQQDExNXaW5kb3dzIEF6dXJlIFRvb2xzMB4XDTE1MDYxNzE3MDAwM1oXDTE1MDYxOTE3MTAwM1owHjEcMBoGA1UEAxMTV2luZG93cyBBenVyZSBUb29sczCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJhjKYhKBkE9TN6WLn8mlvkg3nCgbYwWJSuyMpcE2NH3B/rxndF1a/Mz3MJ5s3+EB9FJpuh5W2ZoqcZ2fpA3nSd+L1zrXY9gOqKqrLmRI3TwSF8X4UVXlA3eFu2QV18FfaNyDE6eBX9Rf0cDTWvNraCW/31fVGs+66d1qDhiEevPo6B9JXXauAN4NgA6uI8rwY/UOQY+3hIVjWJ/FZOAgN/T6Jb/hglCqVcyBn5bxJa3Zz49P3YSBwauf2uyiJlU64B3xmrT75Qo/+cRk5BpkWsBh57PxZhfgwL6IXSezcGTEBl5S3LEMeo+T0To6d+f/F38g2PLqteWHrtTWu+SGYkCAwEAAaMXMBUwEwYDVR0lBAwwCgYIKwYBBQUHAwIwDQYJKoZIhvcNAQEFBQADggEBAEspmwyUvBmbX97gwA3mMFiHyZXFL5MgcA0qjZWy7tUS3gcrLZRSv8SMmX8Yv9F4wMy1oe/H4PxCM2IX8cuQLZ2NfAo2sGVlmhdP12ksW/rKimHpnUTfSnslzqFg4sExu4WIkRn6AJMeqAjOul+A1a6UWKa+RRoRQc6MU/tvBAiEnlRAzFpv/4Ri3yS+KXD0gkOMpoJABkOEtdrgbyFICjfsaku4q1q8LycusFpbE+GEL5TWbUG9wwLZl/IPoy8NBtrv/7Dz8t8a6j/LUxvcChNSj4sID98uasBzOJ8I65xixv+XZXl76a0Q2J4icGkmomktz3TCd7yTVykjgyAbP8c=\"\r\n  }\r\n}",
+      "RequestHeaders": {
+        "Content-Type": [
+          "application/json"
+        ],
+        "Content-Length": [
+          "1035"
+        ],
+        "Accept-Language": [
+          "en-us"
+        ],
+        "x-ms-version": [
+          "2013-03-01"
+        ],
+        "User-Agent": [
+          "Microsoft.Azure.Management.BackupServices.BackupServicesManagementClient/0.0.0.0"
+        ]
+      },
+      "ResponseBody": "{\r\n  \"properties\": {\r\n    \"certificate\": \"MIIC3TCCAcWgAwIBAgIQaa4WR+3Ra7pPjIzudnuuQTANBgkqhkiG9w0BAQUFADAeMRwwGgYDVQQDExNXaW5kb3dzIEF6dXJlIFRvb2xzMB4XDTE1MDYxNzE3MDAwM1oXDTE1MDYxOTE3MTAwM1owHjEcMBoGA1UEAxMTV2luZG93cyBBenVyZSBUb29sczCCASIwDQYJKoZIhvcNAQEBBQADggEPADCCAQoCggEBAJhjKYhKBkE9TN6WLn8mlvkg3nCgbYwWJSuyMpcE2NH3B/rxndF1a/Mz3MJ5s3+EB9FJpuh5W2ZoqcZ2fpA3nSd+L1zrXY9gOqKqrLmRI3TwSF8X4UVXlA3eFu2QV18FfaNyDE6eBX9Rf0cDTWvNraCW/31fVGs+66d1qDhiEevPo6B9JXXauAN4NgA6uI8rwY/UOQY+3hIVjWJ/FZOAgN/T6Jb/hglCqVcyBn5bxJa3Zz49P3YSBwauf2uyiJlU64B3xmrT75Qo/+cRk5BpkWsBh57PxZhfgwL6IXSezcGTEBl5S3LEMeo+T0To6d+f/F38g2PLqteWHrtTWu+SGYkCAwEAAaMXMBUwEwYDVR0lBAwwCgYIKwYBBQUHAwIwDQYJKoZIhvcNAQEFBQADggEBAEspmwyUvBmbX97gwA3mMFiHyZXFL5MgcA0qjZWy7tUS3gcrLZRSv8SMmX8Yv9F4wMy1oe/H4PxCM2IX8cuQLZ2NfAo2sGVlmhdP12ksW/rKimHpnUTfSnslzqFg4sExu4WIkRn6AJMeqAjOul+A1a6UWKa+RRoRQc6MU/tvBAiEnlRAzFpv/4Ri3yS+KXD0gkOMpoJABkOEtdrgbyFICjfsaku4q1q8LycusFpbE+GEL5TWbUG9wwLZl/IPoy8NBtrv/7Dz8t8a6j/LUxvcChNSj4sID98uasBzOJ8I65xixv+XZXl76a0Q2J4icGkmomktz3TCd7yTVykjgyAbP8c=\",\r\n    \"resourceId\": 3420983535961922368,\r\n    \"globalAcsNamespace\": \"seadev01rp1users\",\r\n    \"globalAcsHostName\": \"accesscontrol.windows.net\",\r\n    \"globalAcsRPRealm\": \"http://windowscloudbackup/m3\",\r\n    \"subject\": \"CN=Windows Azure Tools\",\r\n    \"validFrom\": \"2015-06-17T22:30:03+05:30\",\r\n    \"validTo\": \"2015-06-19T22:40:03+05:30\",\r\n    \"thumbprint\": \"39D0915AC95DFD234575A54502AEBC78E344B3E8\",\r\n    \"friendlyName\": \"\",\r\n    \"issuer\": \"CN=Windows Azure Tools\"\r\n  }\r\n}",
+      "ResponseHeaders": {
+        "Content-Length": [
+          "1409"
+        ],
+        "Content-Type": [
+          "application/json"
+        ],
+        "Expires": [
+          "-1"
+        ],
+        "Pragma": [
+          "no-cache"
+        ],
+        "x-ms-request-id": [
+          "c66c0373-80b0-43ff-8089-8075e57913e9"
+        ],
+        "x-ms-client-request-id": [
+          "773dc534-8d14-4812-912f-d2834a9e8c05",
+          "773dc534-8d14-4812-912f-d2834a9e8c05"
+        ],
+        "Strict-Transport-Security": [
+          "max-age=31536000; includeSubDomains"
+        ],
+        "x-ms-ratelimit-remaining-subscription-writes": [
+          "1198"
+        ],
+        "x-ms-correlation-request-id": [
+          "c66c0373-80b0-43ff-8089-8075e57913e9"
+        ],
+        "x-ms-routing-request-id": [
+          "CENTRALUS:20150617T171018Z:c66c0373-80b0-43ff-8089-8075e57913e9"
+        ],
+        "Cache-Control": [
+          "no-cache"
+        ],
+        "Date": [
+          "Wed, 17 Jun 2015 17:10:17 GMT"
+        ],
+        "Server": [
+          "Microsoft-IIS/8.0"
+        ]
+      },
+      "StatusCode": 200
+    }
+  ],
+  "Names": {},
+  "Variables": {
+    "SubscriptionId": "f5303a0b-fae4-4cdb-b44d-0e4c032dde26"
+  }
+}
